streamlit_airtable/requirements.txt
streamlit
torch
transformers
airtable-python-wrapper